Quite a refactor to access default templates from builtin commands
I had to extract the command interface up so it can be accessed from
multiple places. I also had to complete the interface with a bunch of
other things that were required on execution and such.

All in all it makes sense and we can record logs, and also get a tail of
the last command executed, but it was quite some work.