Improve the help command, quite a lot
Now when generating a command, help can be split between the summary and
a list of arguments.

When invoking help it will print the list of commands, but if it is
invoked with a command name, it will print the summary and the list of
arguments, making the output much more explicit.