Still advancing, but things start to make some sense
I need to change the whole main file to create listeners that don't get
messages but get requests, the message itself has to be wrapped around
the request so i can get a RemoteRequestListener, a SlackRequestListener
and probably an APIRequestListener. This makes much more sense, so
there's a refactor coming this way.

For now, I've something that makes some sense from the client side
perspective, but I need to get the logger the sooner the better.

I will also need to isolate the command termination into an injectable
object.